crud completo

+++ src/controllers/motosController.js
@@ -1,12 +1,16 @@
 const motosRepository = require('../repositories/motosRepository');
 
-const getAllMotos = (req, res) => {
-  const motos = motosRepository.getAllMotos();
-  res.render('motos/index', { motos });
+const getAllMotos = async (req, res) => {
+
+  const motos = await motosRepository.getAllMotos();
+  return res.render('motos/index', { motos });
 };
 
-const getMotoById = (req, res) => {
-  const moto = motosRepository.getMotoById(req.params.id);
+const getMotoById = async (req, res) => {
+  if (!req.session.isAuthenticated)   
+    return res.render('auth/login');
+
+  const moto = await motosRepository.getMotoById(req.params.id);
   if (moto) {
     res.render('motos/show', { moto });
   } else {
@@ -14,17 +18,23 @@ const getMotoById = (req, res) => {
   }
 };
 
-const createMotoForm = (req, res) => {
+const createMotoForm = async (req, res) => {
+  if (!req.session.isAuthenticated)   
+    return res.render('auth/login');
   res.render('motos/create');
 };
 
-const createMoto = (req, res) => {
-  const nuevaMoto = motosRepository.createMoto(req.body);
-  res.redirect('/api/motos');
+const createMoto = async (req, res) => {
+  if (!req.session.isAuthenticated)
+    return res.render('auth/login');
+  const nuevaMoto = await motosRepository.createMoto(req.body);
+  res.redirect('/motos');
 };
 
-const updateMotoForm = (req, res) => {
-  const moto = motosRepository.getMotoById(req.params.id);
+const updateMotoForm = async (req, res) => {
+  if (!req.session.isAuthenticated)   
+    return res.render('auth/login');
+  const moto = await motosRepository.getMotoById(req.params.id);
   if (moto) {
     res.render('motos/edit', { moto });
   } else {
@@ -32,18 +42,22 @@ const updateMotoForm = (req, res) => {
   }
 };
 
-const updateMoto = (req, res) => {
-  const updatedMoto = motosRepository.updateMoto(req.params.id, req.body);
+const updateMoto = async (req, res) => {
+  if (!req.session.isAuthenticated)   
+    return res.render('auth/login');
+  const updatedMoto = await motosRepository.updateMoto(req.params.id, req.body);
   if (updatedMoto) {
-    res.redirect('/api/motos');
+    res.redirect('/motos');
   } else {
     res.status(404).send('Moto no encontrada');
   }
 };
 
-const deleteMoto = (req, res) => {
-  motosRepository.deleteMoto(req.params.id);
-  res.redirect('/api/motos');
+const deleteMoto = async (req, res) => {
+  if (!req.session.isAuthenticated)   
+    return res.render('auth/login');
+  await motosRepository.deleteMoto(req.params.id);
+  res.redirect('/motos');
 };
 
 module.exports = {
